Refactor layout variants to use CVA within the lib utility directory (#89)
