+++ src/layouts/Button.tsx
@@ -1,33 +1,26 @@
 import React from "react"
 import { cn } from "@/lib/utils"
-import { variants } from "@/styles/variants"
+import { buttonVariants } from "@/lib/variants"
+import { type VariantProps } from "class-variance-authority"
 import { Box, BaseProps } from "./Box"
 
-interface ButtonProps extends BaseProps, React.ButtonHTMLAttributes<HTMLButtonElement> {
-  variant?: keyof typeof variants.emphasis
-  intent?: keyof typeof variants.intent
-  size?: "sm" | "md" | "lg"
-  fullWidth?: boolean
+interface ButtonProps
+  extends BaseProps,
+    Omit<React.ButtonHTMLAttributes<HTMLButtonElement>, 'color'>,
+    VariantProps<typeof buttonVariants> {
+  as?: any
+  href?: string
   loading?: boolean
 }
 
 export const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
-  ({ className, variant = "solid", intent = "default", size = "md", fullWidth, loading, children, ...props }, ref) => {
+  ({ className, as = "button", variant, intent, size, fullWidth, loading, children, ...props }, ref) => {
     return (
       <Box
-        as="button"
+        as={as}
         ref={ref as any}
         cursor="pointer"
-        className={cn(
-          "inline-flex items-center justify-center transition-all duration-200 font-bold uppercase tracking-widest disabled:opacity-50 disabled:cursor-not-allowed min-h-[44px] min-w-[44px]",
-          variants.emphasis[variant],
-          variants.intent[intent],
-          size === "sm" && "px-4 py-2 text-[10px]",
-          size === "md" && "px-6 py-3 text-xs",
-          size === "lg" && "px-8 py-4 text-sm",
-          fullWidth && "w-full",
-          className
-        )}
+        className={cn(buttonVariants({ variant, intent, size, fullWidth }), className)}
         {...props}
       >
         {children}
